The long awaited event commit is here, partly.
A new system of transfering data between game objects is now
implemented. It's called and contained in the event file.

When the game objects are updated, they can generate an event.
These events will be acted on at the end of the update cycle.

For now two event types are activee, word_input and word_missed.